requirements.txt
# AWS SDK and Services
boto3>=1.34.0
botocore>=1.34.0

# LangChain with AWS Bedrock support
# Using versions compatible with Python 3.13 and numpy 2.x
langchain-core>=0.3.15
langchain-aws>=0.2.4

# Configuration and utilities
python-dotenv==1.0.1
pydantic>=2.0.0

# REST API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Optional: For testing
requests>=2.31.0
httpx>=0.28.0

